Move scraping to DataController
Also a few first tries at scraping the menu data. It's getting
somewhere, but very tedious. I'm seriously contemplating moving the
scraping part of this to a server that can then also happily cache this
or provide the data to something else as well.

Definitely not going to server-side scraping for the transactions
though. That's just an invitation for a lot of hurt.