now parsing main Speiseplan page instead of feed
This is now the third (or fourth?) time I've completely revamped
this code. OpenMensa looked great at first, then sucked. The RSS
feed looked great at first, then sucked. And now this looks pretty
good, but there are sucky parts in the implementation :D
But since I know of no other ways to get the data (besides an
official API of course), this has got to be it.